+++ flaky/multiprocess_string_io.py
@@ -7,9 +7,9 @@ import multiprocessing
 
 class MultiprocessingStringIO(object):
     """
-    Provide an interface to the multiprocessing ListProxy. The
-    multiprocessing ListProxy is a global object that is instantiated before
-    this class would be called, so this is an object class.
+    Provide a StringIO-like interface to the multiprocessing ListProxy. The
+    multiprocessing ListProxy needs to be instantiated before the flaky plugin
+    is configured, so the list is created as a class variable.
     """
 
     _manager = multiprocessing.Manager()
@@ -23,20 +23,22 @@ class MultiprocessingStringIO(object):
 
     def getvalue(self):
         """
-        Shadow the StringIO method
+        Shadow the StringIO.getvalue method.
         """
         return ''.join(i for i in self.proxy)
 
     def writelines(self, content_list):
         """
-        Shadow the StringIO method. Ingests a list and
+        Shadow the StringIO.writelines method. Ingests a list and
         translates that to a string
         """
-        # every time we see a "\n" we should make a new list item
 
         for item in content_list:
             self.write(item)
 
     def write(self, content):
+        """
+        Shadow the StringIO.write method.
+        """
         content.strip('\n')
         self.proxy.append(content)